Updated the `scorer_class` to `scoring_class` in the load_players_table_day2.sql
Seems like the `scoring_class` we created in day 1 lab is referred here as `scorer_class` which raises an error and doesn't let you insert data into the `players` table (once trying to drop and recreate the table in the beginning of lab 2 video). The drop and recreating the `players` table steps are fine but then when you try to input data into the table with this script it fails because of the `scorer_class` that does not exist in the definition of the players table as a column.